feature(OrganizationApplicationUnitTestSuite.java) add unit test suite for the service and mapper tests
