+++ controller-test/src/test/java/org/restful/test/instances/controller/test/controller/OrganizationApplicationUnitTestSuite.java
@@ -13,7 +13,7 @@ import org.restful.test.instances.commons.categories.UnitTest;
 @IncludeCategory(UnitTest.class)
 @ExcludeCategory(IntegrationTest.class)
 @SuiteClasses({
-        OrganizationControllerUnitTests.class
+        OrganizationControllerUnitTest.class
 })
 @Slf4j
 public class OrganizationApplicationUnitTestSuite {
