+++ controller-test/src/test/java/org/restful/test/instances/controller/test/controller/OrganizationControllerIntegrationTest.java
@@ -48,7 +48,7 @@ import static org.hibernate.validator.internal.util.Contracts.assertTrue;
 @FieldDefaults(
         level = AccessLevel.PRIVATE
 )
-public class OrganizationControllerIntegrationTests {
+public class OrganizationControllerIntegrationTest {
 
     /**
      * The Port.
