+++ db-embedded-test/src/test/java/org/restful/test/instances/db/embedded/test/OrganizationApplicationIntegrationTestSuite.java
@@ -17,7 +17,7 @@ import org.restful.test.instances.commons.categories.UnitTest;
         IntegrationTestOnReal.class
 })
 @SuiteClasses({
-        OrganizationRepositoryTests.class
+        OrganizationRepositoryTest.class
 })
 @Slf4j
 public class OrganizationApplicationIntegrationTestSuite {
