+++ db-testcontainers-test/src/test/java/org/restful/test/instances/db/testcontainers/test/OrganizationApplicationIntegrationTestSuite.java
@@ -17,7 +17,7 @@ import org.restful.test.instances.commons.categories.UnitTest;
         IntegrationTestOnEmbedded.class
 })
 @SuiteClasses({
-        OrganizationRepositoryTests.class
+        OrganizationRepositoryTest.class
 })
 @Slf4j
 public class OrganizationApplicationIntegrationTestSuite {
