+++ service-test/src/test/java/org/restful/test/instances/service/test/service/OrganizationApplicationUnitTestSuite.java
@@ -1,4 +1,4 @@
-package org.restful.test.instances.db.local.test;
+package org.restful.test.instances.service.test.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.junit.experimental.categories.Categories;
@@ -6,21 +6,19 @@ import org.junit.experimental.categories.Categories.ExcludeCategory;
 import org.junit.experimental.categories.Categories.IncludeCategory;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite.SuiteClasses;
-import org.restful.test.instances.commons.categories.IntegrationTestOnEmbedded;
-import org.restful.test.instances.commons.categories.IntegrationTestOnReal;
+import org.restful.test.instances.commons.categories.IntegrationTest;
 import org.restful.test.instances.commons.categories.UnitTest;
+import org.restful.test.instances.service.test.service.mapping.OrganizationModelMapperTest;
 
 @RunWith(Categories.class)
-@IncludeCategory(IntegrationTestOnReal.class)
-@ExcludeCategory({
-        UnitTest.class,
-        IntegrationTestOnEmbedded.class
-})
+@IncludeCategory(UnitTest.class)
+@ExcludeCategory(IntegrationTest.class)
 @SuiteClasses({
-        OrganizationRepositoryTest.class
+        OrganizationModelMapperTest.class,
+        OrganizationServiceTest.class
 })
 @Slf4j
-public class OrganizationApplicationIntegrationTestSuite {
+public class OrganizationApplicationUnitTestSuite {
 
     public void main() {
         log.info("Started application tests on emulated database");
